Moved single_threaded code back to openhcl_boot.
