+++ openhcl/openhcl_boot/src/boot_logger.rs
@@ -10,6 +10,7 @@
 
 #[cfg(target_arch = "x86_64")]
 use crate::arch::tdx::TdxIoAccess;
+use crate::single_threaded::SingleThreaded;
 use core::cell::RefCell;
 use core::fmt;
 use core::fmt::Write;
@@ -17,7 +18,6 @@ use core::fmt::Write;
 use minimal_rt::arch::InstrIoAccess;
 use minimal_rt::arch::Serial;
 use minimal_rt::isolation::IsolationType;
-use minimal_rt::single_threaded::SingleThreaded;
 
 /// The logging type to use.
 #[derive(Debug, Clone, Copy, PartialEq, Eq)]
