+++ openhcl/openhcl_boot/src/sidecar.rs
@@ -6,10 +6,10 @@ use crate::host_params::shim_params::ShimParams;
 use crate::host_params::PartitionInfo;
 use crate::host_params::MAX_CPU_COUNT;
 use crate::host_params::MAX_NUMA_NODES;
+use crate::single_threaded::off_stack;
 use arrayvec::ArrayVec;
 use memory_range::MemoryRange;
 use minimal_rt::isolation::IsolationType;
-use minimal_rt::off_stack;
 use sidecar_defs::SidecarNodeOutput;
 use sidecar_defs::SidecarNodeParams;
 use sidecar_defs::SidecarOutput;
